+++ spec/generators/component/blog_generator_spec.rb
@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'rails_helper'
 
 require 'generator_spec/test_case'
 require 'generators/cms/component/blog/blog_generator'
@@ -30,7 +30,7 @@ describe Cms::Generators::Component::BlogGenerator do
   end
 
   it 'create files' do
-    destination_root.should have_structure {
+    expect(destination_root).to have_structure {
       directory 'app' do
         directory 'models' do
           file 'blog.rb' do