+++ spec/generators/component/redirect_generator_spec.rb
@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'rails_helper'
 
 require 'generator_spec/test_case'
 require 'generators/cms/component/redirect/redirect_generator.rb'
@@ -14,7 +14,7 @@ describe Cms::Generators::Component::RedirectGenerator do
   end
 
   it 'creates files' do
-    destination_root.should have_structure {
+    expect(destination_root).to have_structure {
       directory 'app' do
         directory 'models' do
           file 'redirect.rb' do