+++ spec/generators/component/sitemap_generator_spec.rb
@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'rails_helper'
 
 require 'generator_spec/test_case'
 require 'generators/cms/component/sitemap/sitemap_generator.rb'
@@ -27,7 +27,7 @@ describe Cms::Generators::Component::SitemapGenerator do
   end
 
   it 'creates file' do
-    destination_root.should have_structure {
+    expect(destination_root).to have_structure {
       directory 'app' do
         directory 'controllers' do
           file 'sitemap_controller.rb'