+++ spec/generators/widget/flickr_generator_spec.rb
@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'rails_helper'
 
 require 'generator_spec/test_case'
 require 'rails/generators/test_case'
@@ -15,7 +15,7 @@ describe Cms::Generators::Widget::FlickrGenerator do
   end
 
   it 'creates files' do
-    destination_root.should have_structure {
+    expect(destination_root).to have_structure {
       directory 'app' do
         directory 'assets' do
           directory 'stylesheets' do